refactor(websocket): rename the runner to WebSocketClientSession
The class is a per-connection client session, not a runner — it holds the
socket, the auth identity, and the handlers the session owns. "Unified" named
a merge that finished long ago.

`UnifiedWebSocketRunner` and `UnifiedWebSocketRunnerOptions` survive as
aliases and `unified-websocket-runner.ts` as a re-export shim, so none of the
~82 referencing files change here. T8 migrates them and deletes both.

Two path-named checks had to follow the file, since neither tsc nor the
package suite notices a check that identifies code by path: the URL-egress
inventory entry (with its row in docs/url-egress-inventory.md) and the SDK v1
execution contract test, which reads the runner source. Both were inverted and
observed failing before being restored.
