+++ docs/url-egress-inventory.md
@@ -83,7 +83,7 @@ Everything here fetches a URL somebody else chose, through the protected fetch.
 | AudioRef → bytes | `packages/audio-nodes/src/lib/audio-wav.ts` | workflow |
 | Model3DRef / ImageRef → bytes | `packages/video-nodes/src/nodes/model3d/utils.ts` | workflow |
 | `nodetool.model3d.RenderToImage` | `packages/video-nodes/src/nodes/model3d/render.ts` | workflow |
-| chat source images | `packages/websocket/src/unified-websocket-runner.ts` | chat client |
+| chat source images | `packages/websocket/src/websocket-client-session.ts` | chat client |
 | workflow output auto-save | `packages/websocket/src/session/asset-autosave.ts` | workflow |
 | asset ref → bytes on export | `packages/websocket/src/lib/asset-export.ts` | chat client |
 | `save_asset` / `view_image` | `packages/agents/src/capabilities/assets.ts` | model |
